Changes to run in batch
input files are now in /data folder (inputIntrTrans3,csv and
frozenEndosomes.csv)
Output is now in a folder with the date in the /output folder
A "parameter" was created that can be used to scan for different
inputIntrTrans files
BIG PROBLEM
Batch was not able to run COPASI files.  An error occurred because the
CopasiJave file was not found in the Copasi/bin folder. . SOLVED I have
to copy the CopasiJave.dll file to the Copasi/bin folder (C:\Program
Files\copasi.org\COPASI 4.30.240\bin).  THIS NEEDS TO BE DONE IN ALL THE
COMPUTERS WHERE BATCH RUNS ARE PERFORMED
PROBLEM in batch runs, the rutine to save an excel file with all the
agents does not work